Remove hard-coded `true` in `getProcessInfo()`.
This removes the hard-coded `true ||` logic in the `getProcessInfo()` method.
I had needed this because I was never working on a JVM whose name didn't
conform to the expected formatting. As such, I had to force the control-flow.
I forgot to remove it.

This shouldn't have caused too much of a problem since the value was random
(and is the fall-back logic, regardless). But, it's less than ideal.

Special thanks to geoffparky for catching that!